restructure file audits by subclass
